feat(nouvelle-nav): Ajoute la synthèse état des lieux dans un onglet du référentiel
