+++ apps/app/app/(authed)/collectivite/[collectiviteId]/(acces-restreint)/referentiel/new/[referentielId]/(overview)/@tabs/synthese/page.tsx
@@ -1,0 +1,13 @@
+import { SyntheseReferentielView } from '@/app/referentiels/synthese/synthese-referentiel.view';
+import { ReferentielId } from '@tet/domain/referentiels';
+
+export default async function Page({
+  params,
+}: {
+  params: Promise<{
+    referentielId: ReferentielId;
+  }>;
+}) {
+  const { referentielId } = await params;
+  return <SyntheseReferentielView referentielId={referentielId} />;
+}
