+++ apps/app/app/(authed)/collectivite/[collectiviteId]/(acces-restreint)/referentiel/new/[referentielId]/(overview)/@tabs/tabs-wrapper.tsx
@@ -37,6 +37,7 @@ export const TabsWrapper = ({ children }: PropsWithChildren) => {
     <Tabs className="grow flex flex-col" size="sm">
       <TabsList className="!justify-start pl-0 flex-nowrap bg-transparent overflow-x-auto">
         <TabsTab href="progression" label="Mesures" />
+        {!isNewReferentiel && <TabsTab href="synthese" label="Synthèse" />}
         {!isNewReferentiel && !isTableView && (
           <>
             <TabsTab href="priorisation" label="Aide à la priorisation" />
