--- apps/app/app/(authed)/collectivite/[collectiviteId]/(acces-restreint)/referentiel/page.tsx
@@ -1,5 +1,0 @@
-import { TableauDeBordShow } from './tableau-de-bord.show';
-
-export default function Page() {
-  return <TableauDeBordShow />;
-}
